Add files via upload
Random forest model initial attempt - note that I did the initial test/training split with an arbitrary percentage, so I'm trying to do a k-fold cross validation to see how well it performs with different test/training splits, but I'm having some code trouble with that. May or may not be useful to us; happy to discuss in class in a sec here!